backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP requests
httpx==0.25.2
requests==2.31.0

# Date and time utilities
python-dateutil==2.8.2

# Geographic calculations
geopy==2.4.0

# Email support
emails==0.6

# Caching (Redis)
redis==5.0.1
aioredis==2.0.1

# Image processing
Pillow==10.1.0

# Environment variables
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Payment gateway integration (example)
razorpay==1.3.0

# WebSocket support for real-time features
websockets==12.0

# Rate limiting
slowapi==0.1.9

# File uploads
aiofiles==23.2.1

# Scheduled tasks
apscheduler==3.10.4

# Data processing
pandas==2.1.3

# Machine Learning (optional - for AI features)
scikit-learn==1.3.2
transformers==4.35.2

# Monitoring and observability
prometheus-client==0.19.0

# Configuration management
dynaconf==3.2.4

# Backup and export
backup-utils==1.0.0